+++ CHANGELOG.md
@@ -20,6 +20,8 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.4.4] - 2026-06-17
+
 ### Added
 
 - **`PCANManager` implements `IAsyncDisposable`.** Background connection